Move remaining request-conversion functions
